server/pyproject.toml
[project]
name = "campus-bridge"
version = "0.1.0"
description = "Campus Bridge API Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "structlog>=25.5.0",
    "uvicorn>=0.40.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi-injectable>=0.1.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "asyncpg>=0.31.0",
    "alembic>=1.18.1",
    "python-jose>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "bcrypt==4.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/campus_bridge"]

[tool.poe.tasks]
dev = "python scripts/dev/server.py"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "poethepoet>=0.40.0",
    "pre-commit>=4.5.1",
    "pylsp-mypy>=0.7.0",
    "python-lsp-black>=2.0.0",
    "python-lsp-server>=1.14.0",
]

[tool.isort]
profile = "black"
